0.0.4b1 Commit
Releasing since it's been quite a lot of time since the last real update, but note this is incredibly unfinished and buggy. Lotta bad bugs with memory. Most action parameters aren't here either (and, in its current state, supports only strings for butter2short and strings/integers for short2butter). Hence, I am classifying this release as a beta (even though every release of Buttermilk is a beta atm...). Lots of things have changed here, including the biggest one being action parameter support. SeaXML has also got an update for this to work. Butter parameter parsing is a separate header and should be easy to use in libbuttermilk, once it's finished of course.